# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "aptpay App API",
    "description": "Find out how your APIs work",
    "license": {
      "name": "tejas"
    }
  },
  "host": "uataptpay.moreyeahs.in",
  "basePath": "/api",
...
  "schemes": ["https", "http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "APIKeyHeader": []
    }
  ],
  "paths": {
...
    "/disbursement/addDisbursementV1": {
      "post": {
        "summary": "Add Bank and Add Card",
        "tags": ["Disbursement"],
        "description": "Add Disbursement, *referenceId should be unique",
        "parameters": [
          {
            "name": "Add Bank",
            "in": "body",
            "description": "Add Bank",
            "schema": {
              "$ref": "#/definitions/addDisbursementV1"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/addDisbursementV1Response"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/addDisbursementV1ErrorResponse"
            }
          }
        }
      }
    },
    "/disbursement/disbursements/{disbursement_id}": {
      "get": {
        "summary": "get disbursement status",
        "tags": ["Disbursement"],
        "description": "get disbursement status",
        "parameters": [
          {
            "name": "disbursement_id",
            "in": "path",
            "schema": {}
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {}
          },
          "200": {
            "description": " status: false ",
            "schema": {}
          }
        }
      }
    },
    "/disbursement/checkAccount": {
      "post": {
        "summary": "check Account",
        "tags": ["Disbursement"],
        "description": "check Account",
        "parameters": [
          {
            "name": "check Account",
            "in": "body",
            "description": "checkAccount",
            "schema": {
              "$ref": "#/definitions/checkAccount"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/checkAccountResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/checkAccountErrorResponse"
            }
          }
        }
      }
    },
    "/disbursement/addDisbursementInstrument": {
      "post": {
        "summary": "check Account",
        "tags": ["Disbursement"],
        "description": "check Account",
        "parameters": [
          {
            "name": "check Account",
            "in": "body",
            "description": "checkAccount",
            "schema": {
              "$ref": "#/definitions/addDisbursementInstrument"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getDisbursementInstrumentErrorResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/checkAccountErrorResponse"
            }
          }
        }
      }
    },
    "/disbursement/disbursement-instrument/{instrumentId}": {
      "get": {
        "summary": "get disbursement status",
        "tags": ["Disbursement"],
        "description": "get disbursement status",
        "parameters": [
          {
            "name": "instrumentId",
            "in": "path",
            "schema": {}
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {}
          },
          "200": {
            "description": " status: false ",
            "schema": {}
          }
        }
      }
    },
    "/disbursement/disbursement-instrument/list/{payeeId}": {
      "get": {
        "summary": "get disbursement status",
        "tags": ["Disbursement"],
        "description": "get disbursement status",
        "parameters": [
          {
            "name": "payeeId",
            "in": "path",
            "schema": {
              "$ref": "#/definitions/getListofDisbursementInstruments"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getListofDisbursementInstrumentsResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getListofDisbursementInstrumentsErrorResponse"
            }
          }
        }
      }
    },
    "/disbursement/deleteInstrument": {
      "post": {
        "summary": "get disbursement status",
        "tags": ["Disbursement"],
        "description": "get disbursement status",
        "parameters": [
          {
            "name": "instrumentId",
            "in": "query",
            "schema": {
              "$ref": "#/definitions/deleteInstrument"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getListofDisbursementInstrumentsResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getListofDisbursementInstrumentsErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "addDisbursementV1": {
      "properties": {
        "payeeId": {
          "type": "number"
        },
        "amount": {
          "type": "number"
        },
        "transactionType": {
          "type": "string"
        },
        "disbursementNumber": {
          "type": "number"
        },
        "currency": {
          "type": "string"
        },
        "expirationDate": {
          "type": "string"
        },
        "bankNumber": {
          "type": "string"
        },
        "branchTransitNumber": {
          "type": "string"
        },
        "accountNumber": {
          "type": "string"
        },
        "instrumentId": {
          "type": "string"
        },
        "program": {
          "type": "number"
        }
      }
    },
    "addDisbursementV1Response": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "properties": {
            "id": {
              "type": "string"
            },
            "status": {
              "type": "string"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "addDisbursementV1ErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "checkAccount": {
      "properties": {
        "amount": {
          "type": "number"
        },
        "disbursementNumber": {
          "type": "number"
        },
        "currency": {
          "type": "string"
        },
        "expirationDate": {
          "type": "string"
        }
      }
    },
    "checkAccountResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "properties": {
            "id": {
              "type": "string"
            },
            "status": {
              "type": "string"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "checkAccountErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "addDisbursementInstrument": {
      "properties": {
        "payeeId": {
          "type": "number"
        },
        "disbursementNumber": {
          "type": "string"
        },
        "expirationDate": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "bankNumber": {
          "type": "string"
        },
        "branchTransitNumber": {
          "type": "string"
        },
        "type": {
          "type": "number"
        }
      }
    },
...
    "deleteInstrument": {
      "properties": {
        "instrumentId": {
          "type": "string"
        }
      }
    },
...
  }
}